Arreglar error y mostar mensaje cuando no hay cx

+++ nautacli/nauta.py
@@ -2,6 +2,7 @@
 
 from pprint import pprint
 from textwrap import dedent
+from datetime import datetime
 
 import subprocess
 import requests
@@ -28,18 +29,13 @@ LOGOUT_URL_FILE = os.path.join(CONFIG_DIR, "logout_url")
 logfile = open(os.path.join(CONFIG_DIR, "connections.log"), "a")
 
 def log(*args, **kwargs):
-    # date = subprocess.check_output("date").decode().strip()
-    # kwargs.update(dict(file=logfile))
-    # print(
-    #     "{:.3f} ({})".format(
-    #         time.time(),
-    #         date,
-    #     ),
-    #     *args,
-    #     **kwargs,
-    # )
-    # logfile.flush()
-    pass
+    kwargs.update(dict(file=logfile))
+    print(
+        "{} ".format(datetime.now()),
+        *args,
+        **kwargs,
+    )
+    logfile.flush()
 
 def get_inputs(form_soup):
     form = {}
@@ -56,6 +52,16 @@ def parse_time(t):
         return h * 3600 + m * 60 + s
     except:
         return 0
+
+def expand_username(username):
+    """If user enters just username (without domain) then expand it"""
+    with dbm.open(CARDS_DB) as cards_db:
+        for user in cards_db:
+            user = user.decode()
+            user_part = user[:user.index('@')]
+            if username.lower() == user_part:
+                return user
+    return username  # not found
 
 def get_password(username):
     with dbm.open(CARDS_DB) as cards_db:
@@ -81,12 +87,8 @@ def select_card():
 
 def up(args):
     """
-    Esto tiene un problema con la navegacion nacional. Cuando estas conectado
-    si intentas llamar a la funcion google.com nunca va a responder, ni tampoco
-    redirecciona al portal de usuario. Asi que es necesario cambiar la forma de
-    determinar si estamos conectados o no.
-    :param args: 
-    :return: 
+    :param args:
+    :return:
     """
 
     if args.username:
@@ -237,16 +239,18 @@ def down(args):
         return
     session = requests.Session()
     print("Logging out...")
+    r = None
     for error_count in range(10):
         try:
             r = session.get(logout_url)
             break
         except requests.RequestException:
-            print("There was a problem logging out, retrying...")
-            continue
-    print(r.text)
-    if 'SUCCESS' in r.text:
-        os.remove(LOGOUT_URL_FILE)
+            print("There was a problem logging out, retrying %d..." % error_count)
+    if r:
+        log("Logout message: %s" % r.text)
+        if 'SUCCESS' in r.text:
+            print('Connection closed successfully')
+            os.remove(LOGOUT_URL_FILE)
 
 def fetch_expire_date(username, password):
     session = requests.Session()
@@ -277,15 +281,14 @@ def time_left(username, fresh=False, cached=False):
         card_info = json.loads(cards_db[username].decode())
         last_update = card_info.get('last_update', 0)
         password = card_info['password']
-        if not cached:
-            if (now - last_update > 60) or fresh:
-                time_left = fetch_usertime(username)
-                last_update = time.time()
-                if re.match(r'[0-9:]+', time_left):
-                    card_info['time_left'] = time_left
-                    card_info['last_update'] = last_update
-                    cards_db[username] = json.dumps(card_info)
-        time_left = card_info.get('time_left', '-')
+        if not cached and (fresh or now - last_update > 60):
+            time_left = fetch_usertime(username)
+            last_update = time.time()
+            if re.match(r'[0-9:]+', time_left):
+                card_info['time_left'] = time_left
+                card_info['last_update'] = last_update
+                cards_db[username] = json.dumps(card_info)
+        time_left = card_info.get('time_left', 'N/A')
         return time_left
 
 def expire_date(username, fresh=False, cached=False):
@@ -295,12 +298,11 @@ def expire_date(username, fresh=False, cached=False):
     # never change in the case of non-rechargeable cards
     with dbm.open(CARDS_DB, "c") as cards_db:
         card_info = json.loads(cards_db[username].decode())
-        if not cached:
-            if (not 'expire_date' in card_info) or fresh:
-                password = card_info['password']
-                exp_date = fetch_expire_date(username, password)
-                card_info['expire_date'] = exp_date
-                cards_db[username] = json.dumps(card_info)
+        if not cached and (fresh or not 'expire_date' in card_info):
+            password = card_info['password']
+            exp_date = fetch_expire_date(username, password)
+            card_info['expire_date'] = exp_date
+            cards_db[username] = json.dumps(card_info)
         exp_date = card_info['expire_date']
         return exp_date
 
@@ -330,13 +332,25 @@ def cards(args):
             if not args.v:
                 password = "*" * len(password)
             entries.append((card, password))
-    
+
+    con_error = False
     for card, password in entries:
-        print("{}\t{}\t{}\t(expires {})".format(
+        if not con_error:
+            try:
+                time = time_left(card, args.fresh, args.cached)
+                expiry = expire_date(card, args.fresh, args.cached)
+            except requests.exceptions.ConnectionError:
+                con_error = True
+                print('WARNING: It seems that you have no network access. Showing data from cache.')
+        
+        if con_error:
+            time = time_left(card, args.fresh, True)
+            expiry = expire_date(card, args.fresh, True)
+
+        print("{}\t{}\t(expires {})".format(
             card,
-            password,
-            time_left(card, args.fresh, args.cached),
-            expire_date(card, args.fresh, args.cached)
+            time,
+            expiry
         ))
 
 def verify(username, password):
@@ -362,7 +376,7 @@ def cards_add(args):
         print("Credentials seem incorrect")
         return
     with dbm.open(CARDS_DB, "c") as cards_db:
-        cards_db[username] = json.dumps({
+        cards_db[username.lower()] = json.dumps({
             'password': password,
         })
 
@@ -422,7 +436,7 @@ def main():
         epilog=dedent("""\
         Subcommands:
 
-          up [username]
+          up [-t] [username]
           down
           cards [-v] [-f] [-c]
           cards add [username]
@@ -479,18 +493,21 @@ def main():
     down_parser.set_defaults(func=down)
 
     args = parser.parse_args()
-    
-    if 'username' in args and '@' not in args.username:
-            # default domain is @nauta.com.cu
-            args.username += '@nauta.com.cu'
 
+    if 'username' in args and args.username and '@' not in args.username:
+        args.username = expand_username(args.username)
+    
     if args.debug:
         logging.basicConfig(level=logging.DEBUG)
         from http.client import HTTPConnection
         HTTPConnection.debuglevel = 2
 
     if 'func' in args:
-        args.func(args)
+        try:
+            args.func(args)
+        except requests.exceptions.ConnectionError as ex:
+            print("Conection error. Check your connection and try again.")
+            import traceback
+            log(traceback.format_exc())
     else:
         parser.print_help()
-
